Creating the 'extrato_filter.py' file
Removing the 'ExtratoFilter' class from the 'extrato_side_bar.py' to the 'extrato_filter.py' file
